requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.1.1
django-cors-headers==3.5.0
django-ratelimit==3.0.1
djangorestframework==3.11.1
gunicorn==20.0.4
nanoid==2.0.0
psycopg2-binary==2.8.6
pytz==2020.1
sentry-sdk==0.17.6
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
